Adds DetachableProp class and a DetachableGasPump example
Attempting to make a base DetachableProp class for props that can be crashed into and broken, like bins and lampposts etc.
Ported the existing gas pump prop over to the base class and got it to play & stop its fire effect and jump into the air a bit when first hit.